moves handbook files into pages
