+++ frontend/src/components/HandbookOverlay/Pages/Glossary.ts
