+++ frontend/src/components/HandbookOverlay/Pages/HandbookAttacks.tsx
